repaired include processes

+++ Dwarpo/BaseCreature.cpp
@@ -9,10 +9,10 @@ void BaseCreature::initSpeed() {
 	int random = rand();
 	if (random % 5 == 0) {
 		if (random % 2 == 0) {
-			speedModifier = random % 3;
+			speedModifier =(float) (random % 3);
 		}
 		else {
-			speedModifier = 1 / (random % 3);
+			speedModifier = (float)(1 / (random % 3));
 		}
 		baseSpeed *= speedModifier;
 		huntingSpeed *= speedModifier;
